+++ code/lib/cli/src/js-package-manager/Yarn2Proxy.ts
@@ -5,6 +5,36 @@ import type { PackageJson } from './PackageJson';
 export class Yarn2Proxy extends JsPackageManager {
   readonly type = 'yarn2';
 
+  installArgs: string[] | undefined;
+
+  getInstallArgs(): string[] {
+    if (!this.installArgs) {
+      this.installArgs = [];
+
+      if (this.detectWorkspaceRoot()) {
+        this.installArgs.push('-W');
+      }
+    }
+    return this.installArgs;
+  }
+
+  detectWorkspaceRoot() {
+    const { workspaces } = this.readPackageJson();
+
+    if (Array.isArray(workspaces)) {
+      return workspaces.length > 0;
+    }
+
+    if (workspaces && typeof workspaces === 'object') {
+      if (workspaces.packages) {
+        return workspaces.packages.length > 0;
+      }
+      return false;
+    }
+
+    return false;
+  }
+
   initPackageJson() {
     return this.executeCommand('yarn', ['init']);
   }
@@ -31,7 +61,7 @@ export class Yarn2Proxy extends JsPackageManager {
   }
 
   protected runInstall(): void {
-    this.executeCommand('yarn', [], 'inherit');
+    this.executeCommand('yarn', ['install', ...this.getInstallArgs()], 'inherit');
   }
 
   protected runAddDeps(dependencies: string[], installAsDevDependencies: boolean): void {
@@ -41,13 +71,13 @@ export class Yarn2Proxy extends JsPackageManager {
       args = ['-D', ...args];
     }
 
-    this.executeCommand('yarn', ['add', ...args], 'inherit');
+    this.executeCommand('yarn', ['add', ...this.getInstallArgs(), ...args], 'inherit');
   }
 
   protected runRemoveDeps(dependencies: string[]): void {
     const args = [...dependencies];
 
-    this.executeCommand('yarn', ['remove', ...args], 'inherit');
+    this.executeCommand('yarn', ['remove', ...this.getInstallArgs(), ...args], 'inherit');
   }
 
   protected runGetVersions<T extends boolean>(